verif si connecté

+++ view/forum/listSujets.php
@@ -25,21 +25,22 @@ if(App\Session::getVisiteur()){ ?>
                     <p class="par">par</p>
                     <p class="pseudo"><?=$sujet->getVisiteur()->getPseudonyme()?></p>
                     <?php
-                    // var_dump($sujet->getVisiteur());
-                    if(App\Session::isAdmin() || App\Session::getVisiteur()->getId() == $sujet->getVisiteur()->getId()){
-                        if($sujet->getStatut() == 0){ ?>
-                            <a class="statut unlocked" href="index.php?ctrl=forum&action=verrouillerSujet&id=<?= $sujet->getId()?>"><i class="fa-solid fa-lock-open"></i>Sujet déverrouillé</a>
+                    if(App\Session::getVisiteur()){
+                        if(App\Session::isAdmin() || App\Session::getVisiteur()->getId() == $sujet->getVisiteur()->getId()){
+                            if($sujet->getStatut() == 0){ ?>
+                                <a class="statut unlocked" href="index.php?ctrl=forum&action=verrouillerSujet&id=<?= $sujet->getId()?>"><i class="fa-solid fa-lock-open"></i>Sujet déverrouillé</a>
+                            <?php
+                            }else{ ?>
+                                <a class="statut locked" href="index.php?ctrl=forum&action=deverrouillerSujet&id=<?= $sujet->getId()?>"><i class="fa-solid fa-lock"></i>Sujet vérouillé</a>
+                            <?php } ?>
+                            <?php
+                        } ?>
                         <?php
-                        }else{ ?>
-                            <a class="statut locked" href="index.php?ctrl=forum&action=deverrouillerSujet&id=<?= $sujet->getId()?>"><i class="fa-solid fa-lock"></i>Sujet vérouillé</a>
-                        <?php } ?>
-                        <?php
-                     } ?>
-                    <?php
-                    if(App\Session::isAdmin()) {
-                    ?>
-                        <a class="supprimer" href="index.php?ctrl=forum&action=supprimerSujet&id=<?= $sujet->getId()?>"><i class="fa-solid fa-trash"></i>Supprimer<a>
-                    <?php } ?>
+                        if(App\Session::isAdmin()) {
+                        ?>
+                            <a class="supprimer" href="index.php?ctrl=forum&action=supprimerSujet&id=<?= $sujet->getId()?>"><i class="fa-solid fa-trash"></i>Supprimer<a>
+                        <?php }
+                    } ?>
                 </div>
             </a>
         <?php
